requirements-dev.txt
coverage==7.10.6
pytest==8.4.2
pytest-django==4.11.1
